Updating Footer Component
